app: remove app/lib, move all into app
